busy waiting i guess

+++ projeto-2/client/src/main.c
@@ -36,9 +36,9 @@ sem_t thread_counter;
 void *th_request(void *arg);
 
 void free_shared_memory() {
-    if (sem_destroy(&thread_counter)) {
+    /*if (sem_destroy(&thread_counter)) {
         error_sys_ignore_alarm("exit", "couldn't destroy thread semaphore", alarm_status);
-    }
+    }*/
 
     if (close(req_fifo)) {
         error_sys("exit", "couldn't close public request FIFO");
@@ -71,26 +71,26 @@ int main(int argc, char *argv[]) {
         return -1;
     }
 
-    sem_t *client_sem;
+    //sem_t *client_sem;
     char req_fifo_path[BUFFER_SIZE];
 
     sprintf(req_fifo_path, "/tmp/%s", info.path);
 
     int exec_secs = info.exec_secs;
 
-    if ((req_fifo = open(req_fifo_path, O_WRONLY | O_NONBLOCK)) == -1) {
+    if ((req_fifo = open(req_fifo_path, O_WRONLY)) == -1) {
         char error[BUFFER_SIZE];
         sprintf(error, "no public request FIFO of name %s", info.path);
         free_parse_info(&info);
         return error_sys(argv[0], error);
     }
 
-    if (init_sync(0)) {
+    /*if (init_sync(0)) {
         char error[BUFFER_SIZE];
         sprintf(error, "couldn't synchronize with server of public channel %s", info.path);
         free_parse_info(&info);
         return error_sys(argv[0], error);
-    }
+    }*/
 
     free_parse_info(&info);
 
@@ -99,7 +99,7 @@ int main(int argc, char *argv[]) {
         return error_sys(argv[0], "couldn't install atexit function");
     }
 
-    sem_init(&thread_counter, SEM_SHARED, 3);
+    //sem_init(&thread_counter, SEM_SHARED, 3);
 
     struct sigaction alarm_action;
     alarm_action.sa_handler = end_requesting;
@@ -125,12 +125,12 @@ int main(int argc, char *argv[]) {
 
     while(requesting) {
 
-        if (sem_wait(&thread_counter)) {
+        /*if (sem_wait(&thread_counter)) {
             if (error_sys_ignore_alarm(argv[0], "error on waiting empty thread slot", alarm_status)) {
                 pthread_exit(0);
             }
             break;
-        }
+        }*/
         int *arg = (int*)malloc(sizeof(int));
         *arg = id_order;
         if (pthread_create(&tid, NULL, th_request, arg)) {
@@ -155,7 +155,7 @@ int main(int argc, char *argv[]) {
             }
         }
         id_order++;
-        usleep(10000);
+        usleep(1000);
     }
 
     // DEV
@@ -166,7 +166,7 @@ int main(int argc, char *argv[]) {
 
 void *th_request(void *arg) {
     if (arg == NULL) {
-        sem_post(&thread_counter);
+        //sem_post(&thread_counter);
         return NULL;
     }
 
@@ -178,67 +178,23 @@ void *th_request(void *arg) {
 
     free(arg);
 
-    if(write_log(&request, "IWANT")) {
-        char program[BUFFER_SIZE];
-        sprintf(program, "request %d", request.id);
-        error_sys(program, "couldn't write log");
-    }
-
     struct sigaction pipe_action;
     pipe_action.sa_handler = SIG_IGN;
     sigemptyset(&pipe_action.sa_mask);
     pipe_action.sa_flags = 0;
     if (sigaction(SIGPIPE, &pipe_action, NULL)) {
-        sem_post(&thread_counter);
+        //sem_post(&thread_counter);
         return NULL;
     }
 
     char reply_fifo_path[BUFFER_SIZE];
     sprintf(reply_fifo_path, "/tmp/%d.%ld", request.pid, request.tid);
 
-    if (mkfifo(reply_fifo_path, 0660)) {
-        close(req_fifo);
+    if(write_log(&request, "IWANT")) {
         char program[BUFFER_SIZE];
         sprintf(program, "request %d", request.id);
-        error_sys(program, "couldn't create private reply FIFO");
-        sem_post(&thread_counter);
-        return NULL;
+        error_sys(program, "couldn't write log");
     }
-
-    int reply_fifo;
-    if ((reply_fifo = open(reply_fifo_path, O_RDONLY | O_NONBLOCK)) == -1) {
-        close(req_fifo);
-        unlink(reply_fifo_path);
-        char program[BUFFER_SIZE];
-        sprintf(program, "request %d", request.id);
-        error_sys(program, "couldn't open private reply FIFO");
-        sem_post(&thread_counter);
-        return NULL;
-    }
-
-    sem_t *sem_reply;
-
-    if ((sem_reply = sem_open_reply(request.pid, request.tid)) == NULL) {
-        close(req_fifo);
-        unlink(reply_fifo_path);
-        char program[BUFFER_SIZE];
-        sprintf(program, "request %d", request.id);
-        error_sys(program, "couldn't open private reply semaphore");
-        sem_post(&thread_counter);
-        return NULL;
-    }
-
-    /*if (sem_wait_send_request()) {
-        char program[BUFFER_SIZE];
-        sprintf(program, "request %d", request.id);
-        error_sys(program, "couldn't open private reply FIFO");
-        close(req_fifo);
-        unlink(reply_fifo_path);
-        sem_free_reply(sem_reply, request.pid, request.tid);
-        sem_post(&thread_counter);
-        return NULL;
-    }*/
-
 
     if (write_request(req_fifo, &request)) {
         char program[BUFFER_SIZE];
@@ -255,23 +211,66 @@ void *th_request(void *arg) {
         } else {
             error_sys(program, "couldn't write to public request FIFO");
         }
-        unlink(reply_fifo_path);
-        sem_post_receive_request();
-        sem_free_reply(sem_reply, request.pid, request.tid);
-        sem_post(&thread_counter);
-        return NULL;
-    }
-    if (sem_post_receive_request()) {
-        unlink(reply_fifo_path);
-        sem_free_reply(sem_reply, request.pid, request.tid);
-        char program[BUFFER_SIZE];
-        sprintf(program, "request %d", request.id);
-        error_sys(program, "couldn't write to public request FIFO");
-        sem_post(&thread_counter);
+        //unlink(reply_fifo_path);
+        //sem_post_receive_request();
+        //sem_free_reply(sem_reply, request.pid, request.tid);
+        //sem_post(&thread_counter);
         return NULL;
     }
 
-    int private_alarm_status = ALARM_CHILL;
+    if (mkfifo(reply_fifo_path, 0660)) {
+        close(req_fifo);
+        char program[BUFFER_SIZE];
+        sprintf(program, "request %d", request.id);
+        error_sys(program, "couldn't create private reply FIFO");
+        //sem_post(&thread_counter);
+        return NULL;
+    }
+
+    int reply_fifo;
+    if ((reply_fifo = open(reply_fifo_path, O_RDONLY)) == -1) {
+        close(req_fifo);
+        unlink(reply_fifo_path);
+        char program[BUFFER_SIZE];
+        sprintf(program, "request %d", request.id);
+        error_sys(program, "couldn't open private reply FIFO");
+        //sem_post(&thread_counter);
+        return NULL;
+    }
+
+    //sem_t *sem_reply;
+
+    /*if ((sem_reply = sem_open_reply(request.pid, request.tid)) == NULL) {
+        close(req_fifo);
+        unlink(reply_fifo_path);
+        char program[BUFFER_SIZE];
+        sprintf(program, "request %d", request.id);
+        error_sys(program, "couldn't open private reply semaphore");
+        //sem_post(&thread_counter);
+        return NULL;
+    }*/
+
+    /*if (sem_wait_send_request()) {
+        char program[BUFFER_SIZE];
+        sprintf(program, "request %d", request.id);
+        error_sys(program, "couldn't open private reply FIFO");
+        close(req_fifo);
+        unlink(reply_fifo_path);
+        sem_free_reply(sem_reply, request.pid, request.tid);
+        sem_post(&thread_counter);
+        return NULL;
+    }*/
+    /*if (sem_post_receive_request()) {
+        unlink(reply_fifo_path);
+        //sem_free_reply(sem_reply, request.pid, request.tid);
+        char program[BUFFER_SIZE];
+        sprintf(program, "request %d", request.id);
+        error_sys(program, "couldn't write to public request FIFO");
+        //sem_post(&thread_counter);
+        return NULL;
+    }*/
+
+    /*int private_alarm_status = ALARM_CHILL;
     void private_alarm(int sig) {
         private_alarm_status = ALARM_TRIGGERED;
     }
@@ -286,7 +285,7 @@ void *th_request(void *arg) {
         error_sys(program, "couldn't install alarm");
     }
 
-    pthread_t alarm_tid;
+    pthread_t alarm_tid;*/
 
     /*if (create_alarm(pthread_self(), REPLY_TOLERANCE, SIGUSR1, &alarm_tid)) {
         char program[BUFFER_SIZE];
@@ -299,7 +298,7 @@ void *th_request(void *arg) {
         return NULL;
     }*/
 
-    if (sem_wait_reply(sem_reply)) {
+    /*if (sem_wait_reply(sem_reply)) {
         //stop_alarm(alarm_tid);
         char program[BUFFER_SIZE];
         sprintf(program, "request %d", request.id);
@@ -313,7 +312,7 @@ void *th_request(void *arg) {
         sem_free_reply(sem_reply, request.pid, request.tid);
         sem_post(&thread_counter);
         return NULL;
-    }
+    }*/
 
     /*if (stop_alarm(alarm_tid)) {
         char program[BUFFER_SIZE];
@@ -327,16 +326,21 @@ void *th_request(void *arg) {
     }*/
 
     request_t reply;
-    if (read_reply(reply_fifo, &reply)) {
+    int ret;
+    if ((ret = read_reply(reply_fifo, &reply))) {
         char program[BUFFER_SIZE];
         sprintf(program, "request %d", request.id);
-        if(write_log(&request, "FAILD")) {
-            error_sys(program, "couldn't write log");
+        if (ret == -1) {
+            error_sys(program, "couldn't read from private FIFO");
+        } else {
+            if(write_log(&request, "FAILD")) {
+                error_sys(program, "couldn't write log");
+            }
         }
         close(reply_fifo);
         unlink(reply_fifo_path);
-        sem_free_reply(sem_reply, request.pid, request.tid);
-        sem_post(&thread_counter);
+        //em_free_reply(sem_reply, request.pid, request.tid);
+        //sem_post(&thread_counter);
         return NULL;
     }
 
@@ -359,11 +363,11 @@ void *th_request(void *arg) {
         }
     }
 
-    if (sem_free_reply(sem_reply, request.pid, request.tid)) {
+    /*if (sem_free_reply(sem_reply, request.pid, request.tid)) {
         char program[BUFFER_SIZE];
         sprintf(program, "reply %d", reply.id);
         error_sys(program, "couldn't free private reply semaphore");
-    }
+    }*/
 
     if (close(reply_fifo)) {
         char program[BUFFER_SIZE];
@@ -376,6 +380,6 @@ void *th_request(void *arg) {
         sprintf(program, "reply %d", reply.id);
         error_sys(program, "couldn't unlink private reply FIFO");
     }
-    sem_post(&thread_counter);
+    //sem_post(&thread_counter);
     return NULL;
 }
